Modularizing the ui
